Renaming structuredgrid.py to camelCase
