+++ armi/reactor/blueprints/gridBlueprint.py
@@ -161,7 +161,7 @@ class GridBlueprint(yamlize.Object):
         class.
 
         Includes a ``construct`` method, which instantiates an instance of one
-        of the subclasses of :py:class:`~armi.reactor.grids.structuredgrid.StructuredGrid`.
+        of the subclasses of :py:class:`~armi.reactor.grids.structuredGrid.StructuredGrid`.
         This is typically called from within :py:meth:`~armi.reactor.blueprints.blockBlueprint.BlockBlueprint.construct`,
         which then also associates the individual components in the block with
         locations specifed in the grid.
